move constraints and joins out of question.rb and Magma::Question
